# Optimized computing long range correction for CustomNonbondedForce