--- a/openmmapi/include/openmm/internal/CustomNonbondedForceImpl.h
+++ b/openmmapi/include/openmm/internal/CustomNonbondedForceImpl.h
@@ -9,7 +9,7 @@
  * Biological Structures at Stanford, funded under the NIH Roadmap for        *
  * Medical Research, grant U54 GM072970. See https://simtk.org.               *
  *                                                                            *
- * Portions copyright (c) 2008-2016 Stanford University and the Authors.      *
+ * Portions copyright (c) 2008-2021 Stanford University and the Authors.      *
  * Authors: Peter Eastman                                                     *
  * Contributors:                                                              *
  *                                                                            *
@@ -35,6 +35,7 @@
 #include "ForceImpl.h"
 #include "openmm/CustomNonbondedForce.h"
 #include "openmm/Kernel.h"
+#include "openmm/internal/ThreadPool.h"
 #include "lepton/CompiledExpression.h"
 #include <utility>
 #include <map>
@@ -48,6 +49,7 @@ namespace OpenMM {
 
 class OPENMM_EXPORT CustomNonbondedForceImpl : public ForceImpl {
 public:
+    class LongRangeCorrectionData;
     CustomNonbondedForceImpl(const CustomNonbondedForce& owner);
     ~CustomNonbondedForceImpl();
     void initialize(ContextImpl& context);
@@ -62,16 +64,33 @@ public:
     std::vector<std::string> getKernelNames();
     void updateParametersInContext(ContextImpl& context);
     /**
+     * Prepare for computing the long range correction.  This function pre-computes anything
+     * that depends only on the Force (such as particle parameters) but not on information in
+     * the Context (such as global parameters).  This allows the coefficient to be updated
+     * more quickly when global parameters change.
+     */
+    static LongRangeCorrectionData prepareLongRangeCorrection(const CustomNonbondedForce& force);
+    /**
      * Compute the coefficient which, when divided by the periodic box volume, gives the
      * long range correction to the energy.  If the Force computes parameter derivatives,
      * also compute the corresponding derivatives of the correction.
      */
-    static void calcLongRangeCorrection(const CustomNonbondedForce& force, const Context& context, double& coefficient, std::vector<double>& derivatives);
+    static void calcLongRangeCorrection(const CustomNonbondedForce& force, LongRangeCorrectionData& data, const Context& context, double& coefficient, std::vector<double>& derivatives, ThreadPool& threads);
 private:
     static double integrateInteraction(Lepton::CompiledExpression& expression, const std::vector<double>& params1, const std::vector<double>& params2,
             const CustomNonbondedForce& force, const Context& context, const std::vector<std::string>& paramNames);
     const CustomNonbondedForce& owner;
     Kernel kernel;
+};
+
+class CustomNonbondedForceImpl::LongRangeCorrectionData {
+public:
+    CustomNonbondedForce::NonbondedMethod method;
+    std::vector<std::vector<double> > classes;
+    std::vector<std::string> paramNames;
+    std::map<std::pair<int, int>, long long int> interactionCount;
+    Lepton::CompiledExpression energyExpression;
+    std::vector<Lepton::CompiledExpression> derivExpressions;
 };
 
 } // namespace OpenMM