Fixes #3229.  This PR includes three different optimizations.

1. Any part of the computation that doesn't depend on context parameters is precomputed and saved.  It only needs to be repeated if you call `updateParametersInContext()`, not if you change a global parameter.
2. The computation of integrals is parallelized across all the cores in your CPU.
3. On the CUDA and OpenCL platforms, the computation is driven by a separate worker thread so it can be done at the same time the GPU is computing the forces.

I can't give a single number for how much faster it is.  That depends both on the details of your system and on the relative speed of your CPU and GPU.  In the best case where computing the coefficient on the CPU takes less time than computing the forces on the GPU, changing a global parameter will now have negligible cost.  In other cases it may still be quite expensive, but a lot less expensive than it was before.