@peastman 2021-09-17
I'm not sure what's going on with that.  But it wouldn't help anyway: this hasn't been merged, so it wouldn't be in the nightly build.  Are you able to build from source?