@zhang-ivy 2021-09-17
Ah right, I forgot that it hadn't been merged. I've spent a couple hours trying to install from source and I'm having trouble. I followed the instructions here, but when I try to import openmm, I get:
```
>>> import openmm
Segmentation fault
```

I noticed all of the CUDA tests (and one of the Reference tests) are failing when I run `make test`:
```
38/450 Test  #38: TestReferenceRandom ..............................   Passed    0.99 sec
        Start  39: TestReferenceSettle
 39/450 Test  #39: TestReferenceSettle ..............................   Passed    0.04 sec
        Start  40: TestReferenceVariableLangevinIntegrator
 40/450 Test  #40: TestReferenceVariableLangevinIntegrator ..........   Passed   21.91 sec
        Start  41: TestReferenceVariableVerletIntegrator
 41/450 Test  #41: TestReferenceVariableVerletIntegrator ............   Passed    4.07 sec
        Start  42: TestReferenceVerletIntegrator
 42/450 Test  #42: TestReferenceVerletIntegrator ....................***Failed    0.53 sec
        Start  43: TestReferenceVirtualSites
 43/450 Test  #43: TestReferenceVirtualSites ........................   Passed    0.05 sec
        Start  44: TestCudaAndersenThermostatSingle
 44/450 Test  #44: TestCudaAndersenThermostatSingle .................***Failed    0.38 sec
        Start  45: TestCudaAndersenThermostatMixed
 45/450 Test  #45: TestCudaAndersenThermostatMixed ..................***Failed    0.27 sec
        Start  46: TestCudaAndersenThermostatDouble
 46/450 Test  #46: TestCudaAndersenThermostatDouble .................***Failed    0.25 sec
```