@peastman 2021-09-17
Try running a test individually:

```
./TestCudaAndersenThermostat
```

Does it give you an error message with more information?