@zhang-ivy 2021-09-17
Here's the error I get:
```
(long-range-peter) bash-4.2$ ./TestCudaAndersenThermostat
exception: Error launching CUDA compiler: 256
In file included from /usr/local/cuda/bin/../targets/x86_64-linux/include/cuda_runtime.h:83,
                 from <command-line>:
/usr/local/cuda/bin/../targets/x86_64-linux/include/crt/host_config.h:129:2: error: #error -- unsupported GNU version! gcc versions later than 7 are not supported!
  129 | #error -- unsupported GNU version! gcc versions later than 7 are not supported!
      |  ^~~~~
```

My gcc version is:
```
gcc (GCC) 9.4.0
Copyright (C) 2019 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

```

Originally, I had `cudatoolkit               10.2.89              h8f6ccaa_8    conda-forge` and was running `module load cuda/10.2`, so I thought that upgrading my cudatoolkit to 11.2 and running `module load cuda/11.3` (we don't have 11.2 on our cluster) and re-running `make install` would fix this. It doesn't -- I'm getting the same error as above