@peastman 2021-09-17
Apparently nvcc doesn't like the version of gcc you have.  That isn't something I've encountered.  If you have an older compiler around you can use, you can set the environment variable `CUDA_HOST_COMPILER` to point to it.

Hopefully this won't impact running actual programs, since it will use the runtime compiler instead of nvcc.