@peastman 2021-09-17
I was hoping you could test it first.  I don't want to merge it until we know this approach will actually work.