@peastman 2021-09-18
Could you try running your code against the version you just built?  Like I said, there's a good chance the errors affecting the test cases won't affect other programs, since it will use the runtime compiler instead of nvcc.