@zhang-ivy 2021-09-18
I don't think I'll be able to test with my current build because I'm seeing this error:
```
>>> import openmm
Segmentation fault
```