@jchodera 2021-09-20
> Now we just need to figure out why no nightly builds have been uploaded for a couple of weeks.

Sigh. Looks like the last build/upload on Sep 3 was successful, but I can't log into Azure to maintain/check the builds.
This would be a good time for you to get @jaimergp to transfer the omnia build account to you, perhaps, since only OpenMM is being built with this machinery.