@jchodera 2021-09-20
For some reason, none of my accounts allow me access to Azure.