@zhang-ivy 2021-09-20
Are the nightly builds not uploading because the build recipe hard codes the version number? We ran into this issue before [here](https://github.com/openmm/openmmforcefields/pull/156#issuecomment-871779460))