@zhang-ivy 2021-09-20
Actually, it seems like the the hard coded version number [here](https://github.com/omnia-md/conda-dev-recipes/blob/master/recipes/openmm/meta.yaml#L17) is correct, so that's not the problem