@zhang-ivy 2021-09-22
@peastman @jchodera I've tested out this PR on a perses-generated hybrid system for RBD:ACE2. (The same system/state files that I attached [here](https://github.com/openmm/openmm/issues/3229#issuecomment-908725041)).
Just as a reminder, the hybrid system handles nonbondeds with a `NonbondedForce` that contains non-alchemical atoms and a `CustomNonbondedForce` that contains alchemical atoms. By default, `getUseDispersionCorrection()` is `True` for the `NonbondedForce` and `getUseLongRangeCorrection` is `False` for the `CustomNonbondedForce`.

Here are the experiments I ran:
- Experiment 1: setUseDispersionCorrection(True) for only the NonbondedForce, but not the CustomNonbondedForce (aka the default). It takes 8.1 seconds to run 100 steps of MD (where I change a global param every step).
- Experiment 2: setLongRangeCorrection(True) for both CustomNonbondedForces.  It takes 36.1 seconds to run 100 steps of MD (where I change a global param every step).
- Experiment 3: setLongRangeCorrection(True) for both CustomNonbondedForces (and with the changes in this PR).  It takes 12.9 seconds to run 100 steps of MD (where I change a global param every step).

It seems that for my system, this PR offers a 5-6 fold speed up for computing the long range correction in the CustomNonbondedForce. 

I can't remember what the desired speed up is, @jchodera  can you let me know if you think this is a sufficient speed up? 
For now, I'll assume that this is enough and proceed with my rest over the protocol implementation.