@zhang-ivy 2021-10-01
@peastman : I'm running these experiments on our hpc, so it depends on which node I get assigned. 

I just ran the experiments again with `GeForce GTX 1080Ti` and `Intel(R) Xeon(R) CPU E5-2680 v3 @ 2.50GHz` (24 cores). I'm not sure if anything else is running at the same time on the CPU -- but I'm guessing its possible, as I'm using `mode=shared`, so I think that a single node can be shared by multiple jobs. Maybe this is the reason for the slow down?
Btw, I'm using CUDA 10.2 here, not the latest version.

Here are the latest experiments, with the CPU/GPU mentioned above:
With optimization (calling `context.getParameter()`): 3.3659 s
With optimization (not calling `context.getParameter()`): 0.664 s
Without optimization (calling `context.getParameter()`): 21.133 s
Without optimization (not calling `context.getParameter()`): 0.646 s

@ijpulidos ran some of these experiments on his local machine (which has an external GPU and a built in GPU):
(External) GPU: GeForce GTX Titan X - CPU: Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
With optimization (calling `context.getParameter()`): 1.59 s
With optimization (not calling `context.getParameter()`): 1.29 s

(Built in) GPU: Quadro T500, CPU: Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
With optimization (calling `context.getParameter()`): 6.02 s
With optimization (not calling `context.getParameter()`): 5.80 s

While Iván's experiments show much slower simulation times that what we had previously seen, there isn't a big difference when comparing the time required with vs without the `context.getParameter()` call. This seems to further suggest that the slowdown I'm seeing is related to the nodes being shared by other jobs.